+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 August 2025
 
-* Make it possible to respawn on death
+* Make it possible to respawn on death.
 * Remove some print-to-file functionality on death (tombstone and character sheet). If I miss them in the future I'll add a command for them instead.
 * Now uses rust name generation logic by default.
+* Starting items are now always identified.
 
 ## 5-alpha524